Seaprate lyric and lyrics utils.
